+++ src/citation_compass/__init__.py
@@ -4,10 +4,11 @@ from .citation import (
     cite_function,
     cite_module,
     get_all_citations,
-    get_all_imports,
     get_used_citations,
     reset_used_citations,
 )
+
+from .import_utils import get_all_imports
 
 __all__ = [
     "cite_function",
