+++ src/citation_compass/citation.py
@@ -5,8 +5,7 @@ import inspect
 import logging
 import sys
 
-from .docstring_utils import (
-    check_for_any_citation_keyword,
+from citation_compass.docstring_utils import (
     extract_citation,
     extract_urls,
 )
@@ -191,46 +190,3 @@ def get_used_citations():
 def reset_used_citations():
     """Reset the list of used citations."""
     CITATION_REGISTRY_USED.clear()
-
-
-def get_all_imports(skip_common=True, use_keywords=False):
-    """Return a list of all imports in the software package.
-
-    Parameters
-    ----------
-    skip_common : bool
-        Whether to skip the common imports, such as the built-in modules.
-    use_keywords : bool
-        Uses a heuristic that checks if the import's docstring contains
-        keywords that may indicate a citation.
-
-    Returns
-    -------
-    imports : list of str
-        A list of all imports in the software package.
-    """
-    imports = []
-    for name, module in sys.modules.items():
-        skip = False
-        if name == "__main__":
-            # Skip the main module
-            skip = True
-        elif hasattr(module, "__spec__") and module.__spec__ is not None:
-            # Skip the built-in modules or ones from the python framework.
-            origin = module.__spec__.origin
-            if origin is None:
-                skip = False
-            elif origin == "built-in":
-                skip = True
-            elif origin == "frozen":
-                skip = True
-            elif "Python.framework" in origin:
-                skip = True
-
-        if not skip_common or not skip:
-            if use_keywords and hasattr(module, "__doc__"):
-                if check_for_any_citation_keyword(module.__doc__):
-                    imports.append(name)
-            else:
-                imports.append(name)
-    return imports
